adicionando componentes de produtos e menus
