+++ src/App.js
@@ -16,7 +16,7 @@ function App() {
 
       <Header />
       <Routes>
-        <Route path="/" element={<Produtos />} />
+        <Route path="/produtos/:id" element={<Produtos />} />
         <Route path="/home" element={<Home />} />
         <Route path="/meuspedidos" element={<MeusPedidos />} />
         {/* categorias */}
