+++ src/components/MenuOrdenar/styles.module.scss
@@ -1,0 +1,9 @@
+.container {
+    max-width: 1240px;
+}
+.conteudo {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    margin: 0 0 0 0px;
+}
