+++ src/components/Produtos/index.jsx
@@ -1,0 +1,8 @@
+
+export default function ProdutoHome() {
+    return (<div>
+
+        teste
+    </div>
+    )
+}
