--- src/pages/Produtos/style.css
@@ -1,1 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Inter&display=swap');
